Fix `TokenExpiryInfoSuite` broken by V2 security model
"Fix" means make sure tests work in V1 context. (Not that they're rewritten/created for V2 security model.)
