+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/TokenExpiryInfoSuite.java
@@ -64,16 +64,8 @@ public class TokenExpiryInfoSuite extends HapiSuite {
     private static final Logger log = LogManager.getLogger(TokenExpiryInfoSuite.class);
     private static final String TOKEN_EXPIRY_CONTRACT = "TokenExpiryContract";
     private static final String AUTO_RENEW_ACCOUNT = "autoRenewAccount";
-    private static final String UPDATED_AUTO_RENEW_ACCOUNT = "updatedAutoRenewAccount";
-    private static final String INVALID_ADDRESS = "0x0000000000000000000000000000000000123456";
-    private static final long DEFAULT_MAX_LIFETIME =
-            Long.parseLong(HapiSpecSetup.getDefaultNodeProps().get("entities.maxLifetime"));
-    public static final long MONTH_IN_SECONDS = 7_000_000L;
     private static final String ADMIN_KEY = TokenKeyType.ADMIN_KEY.name();
-    public static final String UPDATE_EXPIRY_INFO_FOR_TOKEN = "updateExpiryInfoForToken";
     public static final String GET_EXPIRY_INFO_FOR_TOKEN = "getExpiryInfoForToken";
-    public static final String UPDATE_EXPIRY_INFO_FOR_TOKEN_AND_READ_LATEST_INFO =
-            "updateExpiryInfoForTokenAndReadLatestInfo";
     public static final int GAS_TO_OFFER = 1_000_000;
 
     public static void main(String... args) {
@@ -93,7 +85,7 @@ public class TokenExpiryInfoSuite extends HapiSuite {
     @Override
     public List<HapiSpec> getSpecsInSuite() {
         return List.of(
-                getExpiryInfoForToken(), updateExpiryInfoForToken(), updateExpiryInfoForTokenAndReadLatestInfo());
+                getExpiryInfoForToken());
     }
 
     private HapiSpec getExpiryInfoForToken() {
@@ -171,208 +163,5 @@ public class TokenExpiryInfoSuite extends HapiSuite {
                                                                                     .getAccountID(AUTO_RENEW_ACCOUNT),
                                                                             THREE_MONTHS_IN_SECONDS)))));
                         }));
-    }
-
-    @SuppressWarnings("java:S5960")
-    private HapiSpec updateExpiryInfoForToken() {
-
-        final AtomicReference<TokenID> vanillaTokenID = new AtomicReference<>();
-        final AtomicReference<AccountID> updatedAutoRenewAccountID = new AtomicReference<>();
-
-        return defaultHapiSpec("UpdateExpiryInfoForToken")
-                .given(
-                        cryptoCreate(TOKEN_TREASURY).balance(0L),
-                        cryptoCreate(AUTO_RENEW_ACCOUNT).balance(0L),
-                        cryptoCreate(UPDATED_AUTO_RENEW_ACCOUNT)
-                                .balance(0L)
-                                .keyShape(SigControl.ED25519_ON)
-                                .exposingCreatedIdTo(updatedAutoRenewAccountID::set),
-                        newKeyNamed(ADMIN_KEY),
-                        uploadInitCode(TOKEN_EXPIRY_CONTRACT),
-                        contractCreate(TOKEN_EXPIRY_CONTRACT).gas(1_000_000L),
-                        tokenCreate(VANILLA_TOKEN)
-                                .supplyType(TokenSupplyType.FINITE)
-                                .treasury(TOKEN_TREASURY)
-                                .expiry(100)
-                                .autoRenewAccount(AUTO_RENEW_ACCOUNT)
-                                .autoRenewPeriod(THREE_MONTHS_IN_SECONDS)
-                                .maxSupply(1000)
-                                .initialSupply(500L)
-                                .adminKey(ADMIN_KEY)
-                                .exposingCreatedIdTo(id -> vanillaTokenID.set(asToken(id))))
-                .when(withOpContext((spec, opLog) -> allRunFor(
-                        spec,
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN,
-                                        asHeadlongAddress(INVALID_ADDRESS),
-                                        DEFAULT_MAX_LIFETIME - 12_345L,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(updatedAutoRenewAccountID.get())),
-                                        MONTH_IN_SECONDS)
-                                .alsoSigningWithFullPrefix(ADMIN_KEY, UPDATED_AUTO_RENEW_ACCOUNT)
-                                .via("invalidTokenTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(vanillaTokenID.get())),
-                                        DEFAULT_MAX_LIFETIME - 12_345L,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(updatedAutoRenewAccountID.get())),
-                                        MONTH_IN_SECONDS)
-                                .via("invalidSignatureTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(vanillaTokenID.get())),
-                                        100L,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(updatedAutoRenewAccountID.get())),
-                                        MONTH_IN_SECONDS)
-                                .alsoSigningWithFullPrefix(ADMIN_KEY, UPDATED_AUTO_RENEW_ACCOUNT)
-                                .via("invalidExpiryTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(vanillaTokenID.get())),
-                                        DEFAULT_MAX_LIFETIME - 12_345L,
-                                        asHeadlongAddress(INVALID_ADDRESS),
-                                        MONTH_IN_SECONDS)
-                                .alsoSigningWithFullPrefix(ADMIN_KEY, UPDATED_AUTO_RENEW_ACCOUNT)
-                                .via("invalidAutoRenewAccountTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(vanillaTokenID.get())),
-                                        DEFAULT_MAX_LIFETIME - 12_345L,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(updatedAutoRenewAccountID.get())),
-                                        1L)
-                                .alsoSigningWithFullPrefix(ADMIN_KEY, UPDATED_AUTO_RENEW_ACCOUNT)
-                                .via("invalidAutoRenewPeriodTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS)
-                                .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(vanillaTokenID.get())),
-                                        DEFAULT_MAX_LIFETIME - 12_345L,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(updatedAutoRenewAccountID.get())),
-                                        MONTH_IN_SECONDS)
-                                .alsoSigningWithFullPrefix(ADMIN_KEY, UPDATED_AUTO_RENEW_ACCOUNT)
-                                .via("updateExpiryTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS))))
-                .then(
-                        childRecordsCheck(
-                                "invalidTokenTxn",
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_TOKEN_ID)),
-                        childRecordsCheck(
-                                "invalidSignatureTxn",
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_SIGNATURE)),
-                        childRecordsCheck(
-                                "invalidExpiryTxn",
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_EXPIRATION_TIME)),
-                        childRecordsCheck(
-                                "invalidAutoRenewAccountTxn",
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_AUTORENEW_ACCOUNT)),
-                        childRecordsCheck(
-                                "invalidAutoRenewPeriodTxn",
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith().status(INVALID_RENEWAL_PERIOD)),
-                        withOpContext((spec, opLog) -> {
-                            final var getTokenInfoQuery = getTokenInfo(VANILLA_TOKEN);
-                            allRunFor(spec, getTokenInfoQuery);
-                            final var expirySecond = getTokenInfoQuery
-                                    .getResponse()
-                                    .getTokenGetInfo()
-                                    .getTokenInfo()
-                                    .getExpiry()
-                                    .getSeconds();
-                            final var autoRenewAccount = getTokenInfoQuery
-                                    .getResponse()
-                                    .getTokenGetInfo()
-                                    .getTokenInfo()
-                                    .getAutoRenewAccount();
-                            final var autoRenewPeriod = getTokenInfoQuery
-                                    .getResponse()
-                                    .getTokenGetInfo()
-                                    .getTokenInfo()
-                                    .getAutoRenewPeriod()
-                                    .getSeconds();
-                            assertEquals(expirySecond, DEFAULT_MAX_LIFETIME - 12_345L);
-                            assertEquals(autoRenewAccount, spec.registry().getAccountID(UPDATED_AUTO_RENEW_ACCOUNT));
-                            assertEquals(autoRenewPeriod, MONTH_IN_SECONDS);
-                        }));
-    }
-
-    private HapiSpec updateExpiryInfoForTokenAndReadLatestInfo() {
-
-        final AtomicReference<TokenID> vanillaTokenID = new AtomicReference<>();
-        final AtomicReference<AccountID> updatedAutoRenewAccountID = new AtomicReference<>();
-
-        return defaultHapiSpec("UpdateExpiryInfoForTokenAndReadLatestInfo")
-                .given(
-                        cryptoCreate(TOKEN_TREASURY).balance(0L),
-                        cryptoCreate(AUTO_RENEW_ACCOUNT).balance(0L),
-                        cryptoCreate(UPDATED_AUTO_RENEW_ACCOUNT)
-                                .keyShape(SigControl.ED25519_ON)
-                                .balance(0L)
-                                .exposingCreatedIdTo(updatedAutoRenewAccountID::set),
-                        newKeyNamed(ADMIN_KEY),
-                        uploadInitCode(TOKEN_EXPIRY_CONTRACT),
-                        contractCreate(TOKEN_EXPIRY_CONTRACT).gas(1_000_000L),
-                        tokenCreate(VANILLA_TOKEN)
-                                .supplyType(TokenSupplyType.FINITE)
-                                .treasury(TOKEN_TREASURY)
-                                .expiry(100)
-                                .autoRenewAccount(AUTO_RENEW_ACCOUNT)
-                                .autoRenewPeriod(THREE_MONTHS_IN_SECONDS)
-                                .maxSupply(1000)
-                                .initialSupply(500L)
-                                .adminKey(ADMIN_KEY)
-                                .exposingCreatedIdTo(id -> vanillaTokenID.set(asToken(id))))
-                .when(withOpContext((spec, opLog) -> allRunFor(
-                        spec,
-                        contractCall(
-                                        TOKEN_EXPIRY_CONTRACT,
-                                        UPDATE_EXPIRY_INFO_FOR_TOKEN_AND_READ_LATEST_INFO,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(vanillaTokenID.get())),
-                                        DEFAULT_MAX_LIFETIME - 12_345L,
-                                        HapiParserUtil.asHeadlongAddress(asAddress(updatedAutoRenewAccountID.get())),
-                                        MONTH_IN_SECONDS)
-                                .alsoSigningWithFullPrefix(ADMIN_KEY, UPDATED_AUTO_RENEW_ACCOUNT)
-                                .via("updateExpiryAndReadLatestInfoTxn")
-                                .gas(GAS_TO_OFFER)
-                                .payingWith(GENESIS))))
-                .then(withOpContext((spec, opLog) -> allRunFor(
-                        spec,
-                        childRecordsCheck(
-                                "updateExpiryAndReadLatestInfoTxn",
-                                SUCCESS,
-                                recordWith().status(SUCCESS),
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .forFunction(FunctionType.HAPI_GET_TOKEN_EXPIRY_INFO)
-                                                        .withStatus(SUCCESS)
-                                                        .withExpiry(
-                                                                DEFAULT_MAX_LIFETIME - 12_345L,
-                                                                updatedAutoRenewAccountID.get(),
-                                                                MONTH_IN_SECONDS)))))));
     }
 }
